Fix customer transaction validation
